Refactoring and adding new FlowRunner and its impl
