+++ marklogic-data-hub/src/main/java/com/marklogic/hub/impl/FlowManagerImpl.java
@@ -24,8 +24,8 @@ import com.marklogic.hub.FlowManager;
 import com.marklogic.hub.HubConfig;
 import com.marklogic.hub.error.DataHubProjectException;
 import com.marklogic.hub.flow.Flow;
-import com.marklogic.hub.flow.FlowImpl;
 import com.marklogic.hub.flow.FlowRunner;
+import com.marklogic.hub.flow.impl.FlowImpl;
 import com.marklogic.hub.flow.impl.FlowRunnerImpl;
 import com.marklogic.hub.step.Step;
 import com.marklogic.hub.util.json.JSONObject;
@@ -193,7 +193,7 @@ public class FlowManagerImpl implements FlowManager {
     }
 
     @Override public FlowRunner newFlowRunner() {
-        return new FlowRunnerImpl(hubConfig);
+        return new FlowRunnerImpl();
     }
 
     private JsonNode flowScaffolding = null;
