+++ marklogic-data-hub/src/main/java/com/marklogic/hub/step/Step.java
@@ -17,6 +17,7 @@
 package com.marklogic.hub.step;
 
 import com.fasterxml.jackson.databind.JsonNode;
+import com.marklogic.hub.step.impl.StepImpl;
 
 import java.util.ArrayList;
 import java.util.Arrays;
